Refactor files so assets are in one folder in src
