+++ src/components/Error/Error.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import './Error.css'
-import pikachu from './sad-pikachu.gif'
+import pikachu from '../../assets/images/sad-pikachu.gif'
+
 
 
 export default function Error() {
