+++ src/components/NotFound/NotFound.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: GeomatrixBold;
-  src: url("../assets/fonts/GeomatrixBold-6YBZ1.ttf");
+  src: url("../../assets/fonts/GeomatrixBold-6YBZ1.ttf");
 }
 
 .NotFound {
